requirements.txt
google-genai
Pillow
fastapi
uvicorn[standard]
opencv-python
scipy
numpy
